# lib/api-zod/src/generated/types/submitTestimonialBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitTestimonialBody {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  role: string;
  /** @minLength 20 */
  quote: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  stars: number;
}